Fix call api in hook folder
