+++ src/features/auth/api.js
@@ -11,3 +11,15 @@ export const RequestApi = {
     return axiosInstance.post("/requests", record);
   },
 };
+
+export const ListeningApi = {
+  getQuestions: () => {
+    return axiosInstance.get("/listening/questions");
+  },
+  submitTest: (answers) => {
+    return axiosInstance.post("/listening/submit", answers);
+  },
+  markQuestion: (questionId, isMarked) => {
+    return axiosInstance.put(`/listening/questions/${questionId}/mark`, { isMarked });
+  }
+};
